Revert changes of M-03
Deploying the portalEnergyToken and bToken contract via the constructor increases the required gas beyond the gas limit.

Changes are reverted to the extend that those token contracts now have to be deployed independently before the Portal is deployed. Then their contract addresses are given to the Portal constructor upon deployment. Then ownership over the token contracts must be transferred to the Portal.