Fix SAY-02
To avoid potential underflows when unstaking, the updateAccount logic has been adjusted.

If the maxLockDuration has been updated since the last user interaction, the user´s account struct will now increase the user´s portalEnergy and maxStakeDebt correctly to make up for the difference in staking conditions. Effectively, the user now maintains the potential to use existing stakes for upfront yield of the current maxLockDuration. Before the changes, existing stakes would only be able to get upfront yield up to the maxLockDuration at the moment of initiating the stake.

The fix not only enhances UX but should also avoid any underflow issues when the maxStakeDebt and portalEnergy is updated from calling unstake() & forceUnstakeAll().