Fix Hats-01-L
set approval to 0 before approving the balance amount in _depositToYieldSource().

This avoids problems with some commonly used tokens like USDT where approving new amounts reverts if the current approval is not 0. Such behaviour would result in failed stake() calls.

While not relevant for this particular Portal (HLP), we implement the fix as a precaution for future Portals as this approval behaviour can easily be overlooked.