Fix Hats-66-H
Exploiters were able to sandwich-attack the internal LP whenever convert() is called and effectively remove most of the value accrual to the LP.

This would have led to an unusable Portal in the near term.

The fix applies a timelock logic on an individual user basis that prevents any address from interacting with the internal LP in quick succession.

The duration of the timelock is set at deployment. If the timelock is a reasonably long duration, like 60 seconds, it can even prevent multi-block MEV.

This restriction should not cause any friction for regular users because trading on the internal LP is not supposed to be a high frequency activity.